refactor(auth): create separate auth layout and enhance login page
- Move login page from (public) to (auth) directory
- Improve login form layout and positioning
