+++ src/app/(auth)/layout.tsx
@@ -1,5 +1,3 @@
-import { Navbar } from "@/components/navbar";
-
 interface Props {
   children?: React.ReactNode;
 }
@@ -7,9 +5,7 @@ interface Props {
 const Layout = ({ children }: Props) => {
   return (
     <div className="flex flex-col min-h-screen">
-      <Navbar />
       <div className="flex-1 bg-[#F4F4F0]">{children}</div>
-      {/* <Footer /> */}
     </div>
   );
 };
